Fix type-checking with mutable type parameters
When checking types that are mutable with a mutable type parameter, we
end up comparing against a placeholder with the parameter as a
requirement. This commit ensures we allow such a comparison, allowing
certain patterns such as taking a `mut Array[T: Compare[T]]` as an
argument and passing it e.g. `mut Array[Int]`.

This fixes https://github.com/inko-lang/inko/issues/579.

Changelog: fixed