Fix detecting project roots for test files
Tests won't reside in ./src but ./test, so we need to handle this when
trying to find a file's project root. To prevent a random path
containing /test/ (e.g. ./src/foo/test/bar.inko) from messing things up,
we only treat the parent directory of /test/ as the project root if it
also contains the src/ directory.