Use ByteArray as input for json.Parser
The JSON parser now takes a `ref ByteArray` as its input, instead of a
`String`. This makes parsing files and the likes a bit more efficient,
as you don't need to first convert the bytes into a `String` just so you
can parse it.

As part of this, json.parse is moved to Json.parse_string, and
Json.parse_bytes is added to parse a `ByteArray`. Finally, the runtime
functions for parsing strings are changed to simply take a pointer and a
size, instead of a string and a range. This removes the need for an
intermediate `String`.

This fixes https://github.com/inko-lang/inko/issues/357.

Changelog: changed