Make types namespace private by default
Namespace private means that a symbol is available to modules residing
in the same root namespace, but not other modules. For example, a
private type defined in `std.foo` is available to `std.foo.bar` and
`std.bar`, but not to `http`, because it doesn't have the same root
namespace (= `std`).

This makes it possible to test private types and methods, provided the
tests reside in the same root namespace (which they do, when following
the testing guidelines). In addition, this makes it possible for modules
in the same root namespace to reuse private code, instead of you being
forced to make such code public.

This fixes https://github.com/inko-lang/inko/issues/581.

Changelog: changed