Rework Int parsing and formatting
Int parsing is once again moved into a single Int.parse method, this
time using an enum to describe the format, and a pure Inko
implementation of the parsing algorithm. Both Int.parse and Float.parse
take a std.iter.Bytes as the input, allowing the use of both String and
ByteArray as the input.

Int formatting is moved into Int.format, taking a similar approach to
Int.parse:

    import std.int.Format

    Int.parse('123', Format.Decimal) # => 123
    123.format(Format.Decimal)       # => "123"

Int.to_string still exists of course, and is just uses
Int.format(Format.Decimal) under the hoods; instead of a runtime library
function.

This commit also introduces std.iter.Peekable, and iterator that allows
peeking of the next value without consuming it. This is used by
Int.parse, removing the need for manually maintaining indexes and
advancing them.

The type std.iter.Bytes (and its implementations) are removed, as this
type isn't necessary: allocating an Option only takes about 10-12
nanoseconds, and in the future we'll optimise this enough to make the
Bytes iterator redundant.

The JSON parser still uses the runtime functions for parsing Int and
Float values, but we'll remove that in a separate commit as this
requires additional changes.

Changelog: changed