Refactor std.json.Parser to use a Bytes/iterator
This allows merging of Json.parse_string and Json.parse_bytes into a
single method, and opens the way to using a Read/streaming type of sorts
as input in the future.

As part of this, the parser now reuses Int.parse and Float.parse,
instead of relying on the Rust runtime functions directly. While this
incurs a bit of extra overhead, it ensures the JSON parser doesn't have
to rely on any internal/unsafe functions.

Changelog: changed