Fix module names for test files in sub directories
When running commands such as `inko check ./std/test/...`, we now
generate the correct module name for any files in the test/ directory,
instead of falling back to the "main" name. This is important, as
otherwise the compiler may incorrectly report certain symbols to be
private, when in fact the importing module resides in the same root
namespace as the symbol's source module.

Changelog: fixed