Add std.io.Buffer
This type supports reading from a Bytes type (String or ByteArray in
practise). This is useful when a method takes a Read type, and you want
to test it without performing actual IO operations, as in this case you
can pass a Buffer as part of your tests.

Changelog: added