Merge Format and DefaultFormatter into Formatter
The original idea behind this split was to allow different formatters.
This however ended up not making much sense. For example, the
Format.write method doesn't require/provide nearly enough information to
meaningfully format objects differently, as all it does is take a String
and store it somewhere.

Instead, Formatter is now a class instead of a trait. To make formatting
easier, we'll introduce helper types (similar to Rust) in a similar
commit. In a future commit we can then hook into those types to make
it easier to present the formatted data in different ways (e.g. with and
without indentation).

Changelog: changed