Add types to make debug formatting easier
This adds the types ArrayFormatter, ObjectFormatter and TupleFormatter,
all defined in the std.fmt module. These types make it easier to format
common objects, such as tuples, enums and regular objects. The API is
largely inspired by how Rust does things (e.g. using its DebugStruct
type).

The use of "finish" exists (similar to Rust) even though I wanted to use
a Drop implementation instead. The need for "finish" exists because in a
call chain (x.tuple(...).field(...)), the last method returns a
reference to the formatter type (e.g. TupleFormatter), but the owned
value it points to is dropped prior to returning that reference,
resulting in a drop error. Using "finish" we can work around that by not
having it return a reference.

Changelog: added