Increment/copy fields when used with a receiver
When returning a field accessed through a receiver, such as
`person.name`, we need to either copy the value (if it's a value type)
or create a reference to it (if it's not). Prior to this commit, we just
kept the value as-is, preventing the increment of a reference count (or
copy), which could then lead to a use-after-free error.

Changelog: fixed