Fix missing drops when ignoring match values
When the first case in a match expression returns a `Nil`, the values
returned by other cases (regardless of their type) are to be ignored,
and dropped at the end of the case body. This wasn't implemented
correctly, resulting in either drop errors (in case a reference was
returned), or a leak (when returning an owned value).

Changelog: fixed