Implement specialization of generics
This commit implements type and method specialization. This removes the
need for runtime pointer tagging and allows unboxed Int and Float
values. In addition, the use of specialization may allow for more
optimizations in the future.

The implementation is a little different from other languages. Instead
of specializing generics over every type assigned to a generic type
parameter, types are grouped into "shapes", and we specialize over
shapes. Int, Float, Bool, Nil, String, and atomic types (channels and
processes) get their own shapes, so we can generate slightly different
code for each.

Other owned values are all grouped into the "owned" shape, while
immutable and mutable references go in the "ref" and "mut" shapes
respectively. This is done because such values are always addressed
through pointers, and thus specializing every type/method over such
types results in many types/methods with virtually identical code.

The nice thing about this setup is that we can provide a better balance
between fast compile times and good runtime performance. In addition, a
future version of the compiler could be extended to dynamically create
shapes for types where this makes sense (e.g. if we at some point decide
to stack allocate certain types).

For generics that are assigned the "owned", "ref" or "mut" shape we use
dynamic dispatch, while for Int, Float, etc we use static dispatch. Our
dynamic dispatch strategy is generally fast enough to permit this, and
it means you don't have to decide between using generics and explicit
dynamic dispatch; instead you use generics most of the time and leave it
up to the compiler to compile the code efficiently.

This does require some restrictions to be put in place. Most notably,
Int, Float and String can no longer be cast to traits, and types are no
longer implicitly compatible with traits and instead require explicit
casts. Int and Float can't be used in dynamic dispatch contexts because
they are unboxed values, and thus there's no header/class we can use for
dynamic dispatch. String isn't allowed because of its use of atomic
reference counting, and supporting that would incur an additional cost
for every dynamic dispatch call site. In practise none of this should be
an issue, as these types are rarely used in dynamic dispatch contexts.

This commit also includes various smaller tweaks to the standard and
runtime library, such as the removal of `std.channel.wait` (because it's
a performance trap), and the ability to use generics with raw pointers
(i.e. `Pointer[T]` and `Pointer[String]` are now valid). These changes
are less interesting though, and were needed in response to/to
facilitate the specialization changes.

This fixes https://github.com/inko-lang/inko/issues/502.

Changelog: changed