Fix float parsing on ARM64 macOS
On macOS, interpreting the data returned by inko_string_to_float() as a
Float directly results in inconsistent results: it works in debug mode,
but not in release mode, unless you print Rust's f64::INFINITY value. To
fix this reliably, we have to funnel the bit pattern through
Float.from_bits to ensure the data is correct and consistent.

Changelog: fixed