Make it easy to write compiler tests in Inko
Writing compiler tests in Rust is rather painful as this involves a lot
of boilerplate code. Due to tests often requiring small changes to the
boilerplate, it's difficult to make this reusable. Even if it were
reusable, Rust in general is quite verbose. More often than not we also
don't really care for the specific state of certain data structures,
instead we just want to assert that source code X produces diagnostics
Y.

In this commit we add a setup that makes it easier to write such
tests. This is done by placing Inko source files in
std/test/diagnostics, and using comments to specify the expected
diagnostics. A regular unit test then parses these files, runs `inko
check` on each file, and compares the produced and expected diagnostics.

For certain tests Rust code may still be required, such as when we want
to assert the type database being in a specific state. Perhaps in the
future we'll apply a similar technique for those cases, but for now the
setup is focused on just compiler diagnostics.