Recovering to owned values when assigning fields
When assigning a process' field, we allow the value type to be owned
provided it originated from a `recover` expression, allowing for code
such as this:

    class async Foo[T] {
      let @field: T
    }

    Foo { @field = recover { ... } }

This way one can take `uni T` values as input but explicitly store them
as owned `T` values.

This commit also fixes a bug where it would be possible to assign
process field values using the syntax `receiver.process = value`, which
shouldn't be allowed.

This fixes https://github.com/inko-lang/inko/issues/641.

Changelog: added