Fix inferring placeholders with ownership
In certain cases a placeholder may require a type to have a compatible
ownership before the type can be assigned to said placeholder. For
example, if a `ref T` is returned and `T` is assigned a placeholder,
then passing the placeholder to `ref Array[Int]` should result in the
placeholder being inferred as `ref Array[Int]`.

Prior to this commit this wasn't the case, and placeholders essentially
lost knowledge of their ownership requirements. In case of the Array
example this would result in the placeholder being inferred as
`Array[Int]`.

This commit fixes this by tracking ownership next to type placeholder
IDs, and by extending the type system (the checker, formatter, etc) to
take this into account. This results in a bit of duplication when
checking for compatible ownerships, but this can't be avoided as in some
cases the checks are in fact slightly different compared to normal
types. For example, passing a `TypeRef::Any(_)` to a `TypeRef::Owned(_)`
isn't allowed, but if a placeholder's ownership requirement is "Any",
then we _do_ allow this, because in the context of placeholders "Any"
means "I don't care" rather than "it may be owned, a ref, or something
else".

This fixes https://github.com/inko-lang/inko/issues/636.

Changelog: fixed