Make type checking return types sound
When a method returns a type parameter or a type containing a type
parameter, without an explicit ownership, we no longer consider
references to the type parameter to be compatible with it. This means
code such as this is no longer valid:

    class Box[V] {
      let @value: V

      fn foo -> V {
        @value
      }
    }

Without these changes the above code is unsound, as for `Box[User]` the
return type of `Box.foo` is inferred as `User`, when in reality a
`ref User` is returned.

Changelog: fixed