Increment references when returning variables
When a scope or method returns a local variable that stores a reference,
we now return a new reference and leave the existing register state
as-is. In some cases this may lead to redundant reference counting
increments, but we should be able to optimize those away in the future
(something we need to do anyway).

This fixes https://github.com/inko-lang/inko/issues/670.

Changelog: fixed