Fix debug info for default trait methods
For default methods, the file path used for debugging information is no
longer fixed to the path of the module that implemented the trait.
Instead, if the default method isn't overwritten the path of the module
that _defines_ the trait is used. If the method _is_ overwritten, the
path of the module that contains the implementation is used. This
ensures debug information uses the file paths that contain the actual
method definition.

Changelog: fixed