Reduce memory usage of expanding instructions
The passes that expand drop and reference instructions split a block's
instruction array into two. Prior to this commit, the capacity of the
initial/head array would remain the same as prior to the split,
resulting in a significant increase in memory usage. This commit fixes
this by shrinking the capacity of said array to the number of
instructions that remain in the block. When compiling Inko's test suite,
this reduces peak memory usage by roughly 280 MiB.

Changelog: performance