Decouple symbol names from type and method IDs
Instead of including type and method IDs in symbol names, we instead use
the shape identifiers. This decouples symbol names (and thus object
files) from the order in which modules are processed. This in turn is
needed for a future object caching setup, as otherwise a simple change
in the order in which files are processed would require invalidating
(many) caches.

See https://github.com/inko-lang/inko/issues/520 for more details.

Changelog: changed