Fix getting core counts
The value reported by Rust's available_parallelism() is affected by
sched_setaffinity(2), which in turn is used by the runtime to pin
threads to cores. The result is that if Inko code spawns another Inko
process, the number of cores it uses is limited to a single core.

To fix this, we now do the following:

1. We reset the affinity before starting the runtime, ensuring child
   processes aren't affected by pinned threads of parent processes.
2. We store the core count in the runtime state and reuse that, ensuring
   we always report the same value for the lifetime of an Inko program.

Changelog: fixed