Remove socket deadline guards
The use of guards seemed like a good idea, but in the end results in a
rather clunky API. For one, socket types need to introduce guard types
to clear the deadline. Second, if a socket is dropped while a guard
still exists, you get a runtime panic.

In the new setup you just set a deadline and have to manually overwrite
or clear it (using Socket.reset_deadline or UnixSocket.reset_deadline).
This isn't a great API, but it's the least invasive and better APIs can
always be built on top. This also comes with the added benefit that the
timeout_after= methods no longer panic.

This fixes https://github.com/inko-lang/inko/issues/358 and
fixes https://github.com/inko-lang/inko/issues/359.