Refactor how guard patterns are compiled
Instead of allocating guard blocks ahead of time and reusing those, the
pattern match compiler is given the HIR expressions of guards directly.
This simplifies the compiler, and ensures that when OR patterns are
guarded, the guards for each branch are compiled separately.

This fixes https://github.com/inko-lang/inko/issues/599.

Changelog: fixed