Fix compiling OR patterns
This fixes the pattern matching compiler to handle patterns such as
`a or a` instead of triggering unreachable code. We also emit warnings
for such code.

While patterns such as these are nonsensical, I don't like the idea of
producing a compile-time error just to work around a broken technique
for expanding OR patterns. The new setup also comes with the benefit of
handling OR patterns in a single place.

This fixes https://github.com/inko-lang/inko/issues/679.

Changelog: fixed