Remove support for escaping whitespace in strings
Inko supported strings like this:

    foo \
      bar \
      baz

Such strings would be parsed as this:

    "foo bar baz"

This feature isn't particularly useful as one can use String.+ instead.
It also complicates automatic code formatting, as the presence of a
backslash isn't retained in the AST. While we could change that, this
would lead to additional complexity that frankly isn't justified.

Changelog: changed