Allow creating pointers to module methods
If a module method is defined using "extern" and has a body, the
compiler guarantees calls of the method use the C calling convention.
This in turn makes it safe/possible to pass pointers to such methods to
C libraries.

As per this commit this isn't terribly useful just yet, as the compiler
still passes the process and state variables as hidden arguments. This
will be solved in a separate commit.

This fixes https://github.com/inko-lang/inko/issues/693.

Changelog: added