Ensure stack sizes are a power of two
Stack sizes can't just be a multiple of the page size, they must also be
a power of two for the stack masking to work. Depending on the
configured stack size this may result in twice the size. To counteract
that somewhat, the default initial stack size is changed from 1 MiB to
512 KiB, resulting in a final stack size of 1 MiB (due to the extra
space needed for guard pages).