Rename methods used for unwrapping values
The renames are as follows:

- "unwrap" is now "get"
- "unwrap_or" is now "or"
- "unwrap_or_else" is now "or_else"
- "expect" is now "or_panic"

These methods are renamed to make the method names a bit shorter, remove
the need for sticking "unwrap" in the names, and because "unwrapping"
isn't really clearly defined in the context of Option and Result types
(especially Result as both Ok and Error can be unwrapped).

This fixes https://github.com/inko-lang/inko/issues/662.

Changelog: changed