Make the import syntax consistent
Import statements suffered from ambiguity when importing module methods,
as statements such as `import a.b` could either mean "import module b
from a" or "import module method b from a". To resolve that, one had to
explicitly use a symbol list, i.e. `import a.(b)`.

This commit changes the import syntax such that:

1. A dot is no longer required between the import path and the symbol
   list
2. The symbol list is now always required

This means that instead of this:

    import foo.bar.SomeType
    import std.process.(panic)
    import bla.(A, B) if foo

You now write this instead:

    import foo.bar (SomeType)
    import std.process (panic)
    import bla (A, B) if foo

This solves the ambiguity issue, keeps the syntax consistent, and makes
it easier to format as part of the upcoming `inko fmt` command.

Changelog: changed