Support automatic formatting of source code
This add the command `inko fmt`, used for formatting (or checking the
formatting of) source files. The command supports formatting both
existing files as well as input read from STDIN.

Style wise, the way closures are formatted is changed: closure arguments
are always passed inside the parentheses, with the purpose of removing
the trailing closure syntax in the release _after_ the one that includes
`inko fmt`. This means that instead of this:

    foo.bar(10, 20) fn { ... }.baz(30)

The code is formatted like so:

    foo.bar(10, 20, fn { ... }).baz(30)

The purpose of this change is to make the syntax more familiar, and to
make both parsing and formatting less complicated. Parsing in particular
is made more difficult by the presence of trailing closures.

This commit also includes changes to string literals: both single and
double quoted strings are now semantically the same, meaning they both
support escape sequences and string interpolation. String interpolation
syntax is also changed from this:

    let name = 'Alice'

    "hello {name}"

To this:

    let name = 'Alice'

    "hello ${name}"

These changes make the syntax more consistent and less ambiguous, and as
a result make formatting the code less of a challenge.

This fixes https://github.com/inko-lang/inko/issues/334.

Changelog: added