Add a few tests for inko fmt
This is not an exhaustive list of tests as we can do most testing by
just formatting the standard library. Instead, the list of tests is
meant to cover edge cases we might not encounter in the standard library
itself.

To ensure `inko fmt --check` in CI doesn't warn for the weird formatting
of the input files, the input/output files are located in std/fixtures,
which is ignored by `inko fmt`.