Support handling of Unix signals
This adds the module std.signal, providing a Signal type used to wait
for the arrival of a Unix signal. For example, one can wait for SIGUSR1
as follows:

    import std.signal (Signal)

    Signal.User1.wait

The implementation uses a dedicated runtime thread for waiting for
signals, while all other threads have all the signals masked to ensure
the signals are sent to the signal handling thread. If no processes are
registered for a signal that's received, its default behaviour is
invoked.

This fixes https://github.com/inko-lang/inko/issues/547.

Changelog: added