Format multi-line conditionals more consistently
When a condition for if/while/etc no longer fits on a single line, the
condition is now placed on a new line and indented, such that instead of
this:

    if foo
      and bar
      and baz
    {
      ...
    }

You end up with this:

    if
      foo and bar and baz
    {
      ...
    }

The previous approach didn't always place the opening curly brace at the
correct location, and depending on the length of the expression would
require multiple lines.

The new style isn't entirely unusual either: Stylua also enforces it,
and languages that require parentheses also use it, it just ends up
looking a little different due to those parentheses:

    if (
      foo and bar and baz
    ) {
      ...
    }

This new style is also applied to build tags for import statements,
along with separating imports with many build tags from any imports that
follow them, as to make them easier to read.