Ensure writes always perform full writes
The runtime uses Rust's Write::write() method for writing to files,
sockets, etc. This method doesn't guarantee to perform a write of all of
the contents, and instead may write fewer bytes in a single call. The
runtime/standard library didn't account for this, resulting in partial
writes.

To fix this, the various inko_X_write_string and inko_X_write_bytes
functions are merged into inko_X_write functions, taking a raw pointer
to the data to write and an integer describing the number of bytes to
write. The standard library then uses these functions to perform full
writes, similar to Rust's Write::write_all() method.

Because Write.write_string and Write.write_bytes now guarantee to write
the entirety of the input, the return type is changed from
`Result[Int, Error]` to `Result[Nil, Error]`, as returning the number of
bytes written is no longer relevant as it's always equal to the input
size.

The implementation doesn't account for signal interrupts. When we
implement https://github.com/inko-lang/inko/issues/547 the plan is to
mask signals such that process threads don't receive interrupts, thereby
removing the need for handling them in various places.

Changelog: fixed