Don't generate goto's for dropping immediates
When "dropping" an immediate value such as an Int, we no longer generate
an explicit goto to connect the current and next basic block. By doing
so, the clean up pass simply removes the entire block if no other
instructions are in the block, preventing us from generating many basic
blocks that simply contain a goto to the next block.

While this doesn't really affect performance, it makes the LLVM IR
smaller and makes it easier to visualize the MIR IR.

Changelog: fixed