Replace the polling crate with rustix code
This removes the use of the "polling" crate for interacting with
epoll/kqueue/etc, in favour of doing so ourselves using rustix. This
reduces the number of dependencies, gives us greater control over how
polling is performed, and ensures we're not impacted by any undesired
upstream changes. For more details, refer to
https://github.com/inko-lang/inko/issues/344#issuecomment-1890274292.

This fixes https://github.com/inko-lang/inko/issues/344.

Changelog: changed