Infer match as nil when any case returns nil
The return type of match expressions is now inferred as Nil when _any_
`case` returns Nil, instead of only doing so when the first case returns
Nil. This removes the need for explicit `nil` expressions if one or more
cases return a `Nil`.

This fixes https://github.com/inko-lang/inko/issues/661.

Changelog: changed