Change drop order for fields and enum cases
The order in which fields and the arguments of enum cases are dropped is
now in reverse-definition order, matching that of local variables. This
should be more natural for types that store references to either
themselves or other values, as this allows you to place those last
instead of first, i.e. something like this:

    class Node[T] {
      @next: Node[T]
      @last: mut Node[T]
    }

This fixes https://github.com/inko-lang/inko/issues/602.

Changelog: changed