Use llvm.read_register for reading rsp
I didn't get this to work in the past, but as it turns out that's due to
me not fully understanding what/how to pass to the metadata to the
read_register() intrinsic. Using this intrinsic makes the code a bit
simpler, and allows compiling of the raw LLVM IR with clang (e.g. for
debugging purposes), as the old approach would result in clang errors.

Changelog: changed