Implement Clone for Map
This allows cloning of a `Map`, provided its keys and values also
implement the `Clone` trait.

This requires a fix to the type checker as to ensure type parameter
bounds are correctly exposed in certain cases. I'm not too fond of the
approach used for fixing this, but any better (long term) fixes likely
require extensive changes or a rewrite of the type checker, something
I'd like to avoid for the time being.

This fixes https://github.com/inko-lang/inko/issues/680.

Changelog: added