Unify the class literal and call syntax
Creating instances of classes is now done using the same syntax as
method calls, so instead of this:

    Person { @name = 'Alice', @age = 42 }

You now write this instead:

    Person(name: 'Alice', age: 42)

Positional arguments are also supported:

    Person('Alice', 42)

This is mainly useful for applying the "newtype" pattern like so:

    class Number {
      let @value: Int
    }

    Number(42)

This change is made as the old syntax introduces ambiguity at the syntax
level. This is best illustrated with the following example:

    if A { ... }

Here it's not clear if this should be parsed as an `if` with `A` as its
condition and `{ ... }` as the body to run, or if it should be parsed as
`if (A { ... })` with a missing body.

The parser still supports the old syntax and the formatter converts this
syntax into the new syntax, making it easier to upgrade projects to this
new syntax.

Changelog: changed