Reduce the amount of type conversions
Various methods that took a generic ToPath or ToString argument are
changed to take a concrete type (e.g. IpAddress) instead. The ToPath
trait in turn has been removed entirely, requiring callers to explicitly
convert values to Path values.

The goal of this change is to reduce the amount of stringly-typed APIs.
In addition, this actually makes certain APIs less prone to error, as
it's now more difficult to pass nonsensical String values to certain
methods (e.g. a String containing an invalid IP address to something
that expects a value that can be converted into an IpAddress).

This fixes https://github.com/inko-lang/inko/issues/632.

Changelog: changed