Fix double dropping nested wildcard values
When matching a value against a wildcard, if a different branch
contained a nested pattern (e.g. `Some(Some(A))`), we'd drop both the
outer-most value and any registers/temporaries used for sub values. This
commit fixes this by ensuring that if in such a case we drop a parent
value, we recursively flag all sub values as moved.

This fixes https://github.com/inko-lang/inko/issues/716.

Changelog: fixed