Remove inko pkg Git tests
These require the presence of a Git repository, meaning they'll fail
when running tests from e.g. a tarball. In addition, these tests can be
quite slow depending on the platform used to run the tests.

Since the tests aren't terribly useful and improving them is a bit of a
pain, we'll just remove them for the time being.