Disambiguate class instances without fields
The syntax `receiver.Type` and `Type` was ambiguous when `Type` refers
to a class, as it's not clear if we're referring to a class or creating
an instance of it. In addition, formatting such code wasn't always
handled correctly.

To resolve this, referring to a class _without_ parentheses means we're
just referring to the type, while the use of parentheses creates an
instance of it. Formatting is changed such that we never remove or add
explicit empty parentheses for such cases.