Fix removing poll registration for socket timeouts
When a socket operation times out, we remove its registration with a
poller (e.g. epoll). When doing so, we must ensure the poller ID stored
in the socket is also removed. If we fail to do so, the next blocking
operation tries to modify a non-existing registration, resulting in a
crash.

Changelog: fixed