Only increase read sizes when beneficial
Read.read_all increases the given read size to reduce the number of
calls to Read.read. This used to be done regardless of the number of
bytes read, which could result in unnecessarily large reads. To prevent
this from happening, we only increase the read size if the actual number
of bytes read equals the desired number of bytes.

Changelog: performance