Optimize String.starts_with/ends_with
Instead of using String.byte_index, these methods use the underlying
logic of the optimized String.== method to more efficiently compare the
head/tail of a String.

In addition, ByteArray.== is implemented in Inko using the same logic as
String.==, removing the need for a runtime library function just to
compare two ByteArray values.

Changelog: performance