Rip out CI path filtering
As it turns out, this doesn't actually work: if job A is skipped due to
a path rule, and job B depends on A, then B isn't run either. To work
around that, one has to resort to various hacks
(https://github.com/orgs/community/discussions/45058), which I'm not
interested in.

Since CI containers change rarely, we'll just move back to using a
separate workflow that we have to trigger in the right order.