Relax LLVM requirements in the install guide
Newer versions of LLVM generally _should_ work with the one we target
(at least now that we target LLVM 17 instead of 16 or even 15), though
it's still preferred to install a package that matches the target
version. In practical terms this means that building Inko on Arch Linux
shouldn't be that difficult after all.